Add SSAR application.
Add SSAR application files from Sourcery VSIPL++ 3.3-12, as provided by J. Kepner.  These files are
licensed under a BSD license; the license headers are exactly as in the files delivered by Mentor Graphics
and have not been altered.

Note that this does not include the SSAR data files for size reasons.  The README file contains pointers to
the origins of this benchmark, and the reference Matlab implementation files (which can be tracked down
from those pointers) contain scripts to regenerate the input data.